[eve MAINLINE] Automatic VCXPROJ submission from 2290
This might create build loops, but it was deemed valuable enough to be able to sync these.
https://teamcity.evetech.net/viewLog.html?buildId=8853704
!nobuild

Source Changes
----------------
[Eve Engine] 1654529 by patrick.kerr (Patrick Kerr)
carbon/src/trinity/RegenerateProjectFile_v141.cmd: Create an explicit v141 ProjectFileGenerator script for trinity, to hopefully "ease the pain" of keeping the v141 project files in-sync with the v100 legacy files, until the transition is finally complete and the v100 stuff can be removed.
----------------

[git-p4: depot-paths = "//eve/branches/development/MAINLINE/carbon/src/audio2/": change = 1654631]