Query RTPCs in a more performant way and add Telemetry insight for Wwises audio thread. (#9)
* Add CPU profiling hooks from Wwise and feed them into Telemetry

Wwise 2022 added hooks that give insight into the CPU usage of the audio
thread. This commit uses those hooks to feed Wwise audio engine CPU
timings into Telemetry. This only works when using the Debug or Profile
flavor of the Wwise SDK however.

* Change methodology for updating/accessing parameter info from curveset drivers

Make it so that AudManager keeps track only of the values of an audio
parameter. Remove the paradigm of setting the value of those audio
parameters in curve set drivers. Change curve set drivers to get their
info about parameters from the audio manager. This was done to make sure
this multi threaded code writes as little as possible.

https://ccpgames.atlassian.net/browse/PLAT-2064