Allow running individual tests (#21)
* Allow running individual tests

Using `run_tests.py` severely limited the way in which the tests could be run. With this change, however, it is possible to run tests individually through CTest, for example. Furthermore, this restructures the repository to be more in line with our reference template project, which in turn avoids publishing the test code into the runtime environment.

* Revert accidental whitespace change

* Use correct pythonInterpreter on each platform