Fix issues with linking against CarbonAudio (#24)
VideoPlayer builds are failing because CarbonAudio links Intel::TBB with PUBLIC linkage, but does not export it as a dependency in its generated config file. Since this dependency can just as well be privately linked, it makes the most sense to do just that.

Also, clean up BlueExposure from the list of linked libraries since we get that transitively when linking Blue.