Make changes to port code to Python3
* Make changes to port code to Python3

* Stop comparing lists of arbitrary order to fix flakey tests and allow tests on macOS

* Correctly target platform specific pythonInterpreter

* Reorder precompiled header includes

As of Python 3 the `<BlueExposure.h>` include pulls in Python 3's `<structmember.h>` which defines a `T_CHAR` symbol. This clashes with a template parameter in the AudioKinetic headers.

* Python3.12 - carbon-audio tests failing due to deprecated functions

Replacing assertEquals with assertEqual

Part of:
https://ccpgames.atlassian.net/browse/PLAT-5682

* Move audio2 imports into local scopes

Otherwise, test discovery of the very first build fails. This issue was masked in other branches because an appropriate _audio2 build was available as part of the client interpreter environment.

* Replace stackless with scheduler

---------

Co-authored-by: Thomas Dähling <thomas@ccpgames.com>
Co-authored-by: Thorsteinn Mar Thorsteinsson <thorsteinnm@ccpgames.com>
Co-authored-by: CCP Zoetrope <eric@ccpgames.com>