Fix defect where having no audio device resulted in log spam (#39)
When a user does not have any audio devices enabled on Windows then it seems like Wwise has some sort of heartbeat mechanism that's always activating their audio device change callback. Because of this, it results in carbon-audio logging an error every 5 seconds or so. To fix this, a hash mechanism is used that hashes all incoming arguments to the audio device callback, keeping it from just doing the same thing over and over until something changes.

https://ccpgames.atlassian.net/browse/PLAT-8062