Remove unnecessary Wwise 2023 linking adjustments from CMakeLists.txt (#55)
Removed patching the Wwise library file directory in the branch from CMakeLists.txt, as it is doing string manipulation on a path string without using the operating system to decide on which separator to use. In addition to this, both the CARBON-AUDIO-DEPLOY EVE sandbox and the FRONTIER-SOUNDBANK-OPTIMISATIONS branch have been updated to have all lib files be in the original location they were in before this change was made.

https://ccpgames.atlassian.net/browse/PLAT-9860