Move logic for stretch effect audio from Trinity to Carbon Audio. (#65)
Move the logic of Tr2AudioStretchBase and Tr2AudioStretchAuto from Trinity to Carbon Audio and:

* Combine the two classes into one, IStretchAudio, because it really didn't make sense to have them separate.
* Make it use the position of the listener instead of the Trinity camera as the guiding light for where to put the stretch audio emitter.
* Simplify the API a bit from what it was in Trinity.

It was also necessary to add a call to Initialize() in the constructor of AudGameObjResource. This is because emitters were not being registered in Wwise when sound prioritization was disabled and an emitter was created in C++. Sound prioritization is not enabled in Graphite which highlighted this problem.

Originally, the intent of this work was to simply fix a bug where the stretch audio emitter was not following the camera (https://ccpgames.atlassian.net/browse/PLAT-10189), but it made more sense to move this logic to Carbon Audio now that we have the carbon-trinityaudioapi. Along with this it is more stable if the audio listener is used as a anchor point for the stretch audio emitter rather than the Trinity camera.


https://ccpgames.atlassian.net/browse/PLAT-10189

* Simplify StretchAudio API and make more readable

* Simplify the StretchAudio API to just have Start and Stop.
* Add private function and extract the logic of projecting the listener position onto a line segment into it.
* Update StretchAudio class description.

https://ccpgames.atlassian.net/browse/PLAT-10189