Add invisible clickable debug rendering sphere  (#77)
Add invisible debug sphere that allows the user to click on an audio emitter in the trinity scene inside Graphite. This debug sphere scales in relation to the distance to the listener.

Before this, the user would have to either painstakingly click the perfect pixel in space or just find the audio emitter in the tree view to be able to select it.

https://ccpgames.atlassian.net/browse/PLAT-10265