Fix defect that stopped all "essential media" from being loaded correctly (#79)
When upgrading to Wwise 2023 the previous logic of using the static data repository to determine if a .wem or .bnk file is essential was accidentally removed. After the 2023 upgrade LowLevelIO has been resolving all .bnk files as essential and all .wem files as non-essential, resulting in any music that depends on essential .wem files to fail to play. This has been rectified.

Add more robust tests for this particular feature to hopefully be able to catch if this breaks again down the line.

https://ccpgames.atlassian.net/browse/PLAT-10139